created polling_engines package & moved relevant code there
