+++ app/src/main/kotlin/com/igor_shaula/api_polling/data_layer/polling_engines/PollingEngine.kt
@@ -1,0 +1,8 @@
+package com.igor_shaula.api_polling.data_layer.polling_engines
+
+interface PollingEngine {
+
+    fun launch(intervalInSeconds: Int, theWorkToDo: () -> Unit)
+
+    fun stopAndClearItself()
+}
